Adding sfpowerscripts as a resource
Hi, I am the author of sfpowerscripts (https://sfpowerscripts.com/),  an extension to azure pipelines which is basically a wrapper around the DX commands and removes the need for scripting it out. I would propose to add this plugin to your guide and also would love to collaborate in this book.